+++ app/services/task/filter/contact_donation_amount_recommendation.rb
@@ -1,19 +1,20 @@
-class Task::Filter::ContactType < Task::Filter::Base
+class Task::Filter::ContactDonationAmountRecommendation < Task::Filter::Base
   def execute_query(tasks, filters)
+    filters = clean_contact_filter(filters)
     # We are plucking ids here because the contact filter already generates an sql statement with several nested subqueries
     # and sending too many of those to postgres can cause unexpected errors and is often slower than breaking things up.
     # Do not change this unless you test the results before pushing to production.
 
     tasks.joins(:contacts)
-         .where(contacts: { id: Contact::Filter::ContactType.query(contact_scope(tasks), filters, account_lists).ids })
+         .where(contacts: { id: Contact::Filter::DonationAmountRecommendation.query(contact_scope(tasks), filters, account_lists).ids })
   end
 
-  def title
-    _('Contact Type')
+  def parent
+    _('Contact Gift Details')
   end
 
   delegate :custom_options,
-           :parent,
            :type,
-           to: 'Contact::Filter::ContactType.new(account_lists)'
+           :title,
+           to: 'Contact::Filter::DonationAmountRecommendation.new(account_lists)'
 end