+++ spec/acceptance/api/v2/contacts/merges_spec.rb
@@ -86,6 +86,7 @@ resource 'Contacts > Merges' do
       contact_referrals_to_me
       contacts_referred_by_me
       contacts_that_referred_me
+      donation_amount_recommendations
       donor_accounts
       last_six_donations
       people